Crate mustache

Re-exports

pub use self::Data::*;

pub use builder::MapBuilder;

pub use builder::VecBuilder;

pub use encoder::Encoder;

pub use encoder::EncoderResult;

Modules

builder

encoder

parser

Structs

Context
Represents the shared metadata needed to compile and render a mustache template.

Template
Template represents a compiled mustache file.

Enums

Data

Error
Error type for any error within this library.

Functions

compile_iter
Compiles a template from an Iterator<char>.

compile_path
Compiles a template from a path. returns None if the file cannot be read OR the file is not UTF-8 encoded

compile_str
Compiles a template from a string.

Type Definitions

Result