Enum mustache::EncoderError

pub enum EncoderError {
    NestedOptions,
    UnsupportedType,
    MissingElements,
    KeyIsNotString,
    NoDataToEncode,
    Message(String),
    // some variants omitted
}

Error type to represent encoding failure.

This type is not intended to be matched exhaustively as new variants may be added in future without a version bump.

Variants
NestedOptionsUnsupportedTypeMissingElementsKeyIsNotStringNoDataToEncodeMessage(String)