Enum mustache::parser::Error

pub enum Error {
    BadClosingTag(char, char),
    UnclosedTag,
    UnclosedSection(String),
    UnbalancedUnescapeTag,
    EmptyTag,
    EarlySectionClose(String),
    MissingSetDelimeterClosingTag,
    InvalidSetDelimeterSyntax,
    // some variants omitted
}

Error type to represent parsing failure.

This type is not intended to be matched exhaustively as new variants may be added in future without a version bump.

Variants
BadClosingTag(char, char)UnclosedTagUnclosedSection(String)UnbalancedUnescapeTagEmptyTagEarlySectionClose(String)MissingSetDelimeterClosingTagInvalidSetDelimeterSyntax