Struct mustache::Context

pub struct Context {
    pub template_path: Path,
    pub template_extension: String,
}

Represents the shared metadata needed to compile and render a mustache template.

Fields

template_path

template_extension

Methods

impl Context

fn new(path: Path) -> Context

Configures a mustache context the specified path to the templates.

fn compile<IT: Iterator<char>>(&self, reader: IT) -> Template

Compiles a template from a string

fn compile_path(&self, path: Path) -> Result<Template, Error>

Compiles a template from a path.