Crate mustache

Reexports

pub use self::Data::*;

pub use builder::{MapBuilder, VecBuilder};

pub use encoder::{Error, IoError, InvalidStr, Encoder, EncoderResult};

Modules

builder

encoder

Structs

Context
Represents the shared metadata needed to compile and render a mustache template.

Template
Template represents a compiled mustache file.

Enums

Data

Functions

compile_iter
Compiles a template from an Iterator<char>.

compile_path
Compiles a template from a path. returns None if the file cannot be read OR the file is not UTF-8 encoded

compile_str
Compiles a template from a string.