Crate mustache

Re-exports

pub use encoder::SerializeVec;

pub use encoder::SerializeTupleVariant;

pub use encoder::SerializeMap;

pub use encoder::SerializeStructVariant;

Structs

Context

Represents the shared metadata needed to compile and render a mustache template.

Encoder

MapBuilder

MapBuilder is a helper type that construct Data types.

Template

Template represents a compiled mustache file.

VecBuilder

Enums

Data

EncoderError

Error type to represent encoding failure.

Error

Error type for any error within this library.

ParserError

Error type to represent parsing failure.

Functions

compile_iter

Compiles a template from an Iterator<char>.

compile_path

Compiles a template from a path. returns None if the file cannot be read OR the file is not UTF-8 encoded

compile_str

Compiles a template from a string.

to_data

Type Definitions

Result