Struct mustache::Context

pub struct Context {
    pub template_path: PathBuf,
    pub template_extension: String,
}

Represents the shared metadata needed to compile and render a mustache template.

Fields
template_path: PathBuf template_extension: String
Methods

impl Context

pub fn new(path: PathBuf) -> Context

Configures a mustache context the specified path to the templates.

pub fn compile<IT: Iterator<Item = char>>(&self, reader: IT) -> Result<Template>

Compiles a template from a string

pub fn compile_path<U: AsRef<Path>>(&self, path: U) -> Result<Template>

Compiles a template from a path.